refactor : 상속 두 번 대신 JpaQueryFactory 주입방식
