+++ 강지민/week8/src/main/java/com/witbackend/week8/repository/MemberRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface MemberRepository extends JpaRepository<Member, Long>, MemberCustomRepository {
+public interface MemberRepository extends JpaRepository<Member, Long> {
 
     List<MemberResponseDto> findAllBy(Pageable pageable);
 }
